+++ relaxed_client/lib/src/relaxed_client_base.dart
@@ -4,4 +4,3 @@
 class Awesome {
   bool get isAwesome => true;
 }
-