+++ relaxed_client/lib/src/status.dart
@@ -4,10 +4,10 @@
 /// These are usually created by the Repository classes where they return
 /// `LiveData<Resource<T>>` to pass back the latest data to the UI with its fetch status.
 enum Status {
-    // ignore: constant_identifier_names
-    SUCCESS,
-    // ignore: constant_identifier_names
-    ERROR,
-    // ignore: constant_identifier_names
-    LOADING
+  // ignore: constant_identifier_names
+  SUCCESS,
+  // ignore: constant_identifier_names
+  ERROR,
+  // ignore: constant_identifier_names
+  LOADING
 }