+++ when_async/lib/src/when_async_base.dart
@@ -1,3 +1,4 @@
+import 'package:meta/meta.dart';
 import 'package:when_async/src/future.dart';
 
 import 'data/data.dart';
@@ -9,9 +10,37 @@ import 'typedefs.dart';
 ///
 /// A common use-case for [When] is to easily consume states of an asynchronous operation such as reading a file or making an HTTP request.
 abstract class When<ASYNC_RESULT_TYPE, RESULT_TYPE> {
-  const When(this.asyncValue);
+  When(AsyncResultBuilderCallback<ASYNC_RESULT_TYPE> builder)
+      : _valueBuilder = builder,
+        _asyncValue = null;
 
-  final ASYNC_RESULT_TYPE asyncValue;
+  When.value(ASYNC_RESULT_TYPE value)
+      : _asyncValue = value,
+        _valueBuilder = _valueBuilderFromValue(value);
+
+  static AsyncResultBuilderCallback<T> _valueBuilderFromValue<T>(T value) {
+    return () => value;
+  }
+
+  ASYNC_RESULT_TYPE? _asyncValue;
+
+  ASYNC_RESULT_TYPE get asyncValue {
+    if (_asyncValue == null) {
+      buildAsyncValue();
+    }
+    return _asyncValue!;
+  }
+
+  final AsyncResultBuilderCallback<ASYNC_RESULT_TYPE> _valueBuilder;
+
+  @protected
+  AsyncResultBuilderCallback<ASYNC_RESULT_TYPE> get valueBuilder =>
+      _valueBuilder;
+
+  @protected
+  void buildAsyncValue() {
+    _asyncValue = valueBuilder();
+  }
 
   /// Asynchronously executes with listener callbacks in parameters for the asynchronous computation that may fail into something that is safe to read.
   ///
@@ -22,6 +51,9 @@ abstract class When<ASYNC_RESULT_TYPE, RESULT_TYPE> {
     final VoidErrorCallback? onError,
     final VoidCallback? onFinally,
   });
+
+  /// Recreates a when instance from the given async builder.
+  When<ASYNC_RESULT_TYPE, RESULT_TYPE> recreate();
 
   /// Asynchronously executes with snapshots callbacks in parameters for the asynchronous computation that may fail into something that is safe to read.
   ///
@@ -34,8 +66,8 @@ abstract class When<ASYNC_RESULT_TYPE, RESULT_TYPE> {
 
   /// A utility that uses [WhenFuture] for executing Futures with snapshots.
   static WhenFuture<RESULT_TYPE> future<RESULT_TYPE>(
-    Future<RESULT_TYPE> asyncValue,
+    AsyncResultBuilderCallback<Future<RESULT_TYPE>> asyncValueBuilder,
   ) {
-    return WhenFuture<RESULT_TYPE>(asyncValue);
+    return WhenFuture<RESULT_TYPE>(asyncValueBuilder);
   }
 }