+++ when_async/CHANGELOG.md
@@ -9,3 +9,8 @@
 ## 1.1.0+1
 
 * Lower dependency constraint of meta to 1.3.0
+
+## 1.2.1
+
+* Add uninitialized future snapshot state
+* Add a async value builder callback for lazy building and recreation of when classes.